立绘参数补丁，float参数，对话主题绑定

+++ src/main/java/com/zhenshiz/chatbox/utils/chatbox/RenderUtil.java
@@ -363,7 +363,7 @@ public class RenderUtil {
     }
 
     // image
-    public static void renderImage(GuiGraphics guiGraphics, ResourceLocation resourceLocation, int x, int y, int z, int uw, int uh, int width, int height) {
+    public static void renderImage(GuiGraphics guiGraphics, ResourceLocation resourceLocation, float x, float y, float z, float uw, float uh, float width, float height) {
         BufferBuilder bufferBuilder = Tesselator.getInstance().begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
         Matrix4f matrix4f = guiGraphics.pose().last().pose();
         bufferBuilder.addVertex(matrix4f, x, y, z).setUv(0, 0);
@@ -377,7 +377,7 @@ public class RenderUtil {
         BufferUploader.drawWithShader(bufferBuilder.buildOrThrow());
     }
 
-    public static void renderImage(GuiGraphics guiGraphics, ResourceLocation resourceLocation, int x, int y, int z, int width, int height, float scale) {
+    public static void renderImage(GuiGraphics guiGraphics, ResourceLocation resourceLocation, float x, float y, float z, float width, float height, float scale) {
         x = (int) (x / scale);
         y = (int) (y / scale);
         guiGraphics.pose().pushPose();
